requirements.txt
dj-database-url==0.4.2
Django==1.10.6
django-tastypie==0.13.3
djangorestframework==3.6.2
gunicorn==19.7.1
python-dateutil==2.6.0
python-mimeparse==1.6.0
requests==2.13.0
six==1.10.0
wheel==0.24.0
whitenoise==3.3.0
psycopg2==2.6.2

